structure moved to ra_ide_api
ra_ide_api_light removed completely
