+++ crates/ra_ide_api/src/structure.rs
