--- crates/ra_ide_api_light/Cargo.toml
@@ -1,26 +1,0 @@
-[package]
-edition = "2018"
-name = "ra_ide_api_light"
-version = "0.1.0"
-authors = ["rust-analyzer developers"]
-publish = false
-
-[dependencies]
-itertools = "0.8.0"
-superslice = "1.0.0"
-join_to_string = "0.1.1"
-rustc-hash = "1.0"
-
-ra_syntax = { path = "../ra_syntax" }
-ra_text_edit = { path = "../ra_text_edit" }
-ra_fmt = { path = "../ra_fmt" }
-
-[dev-dependencies]
-test_utils = { path = "../test_utils" }
-insta = "0.7.0"
-
-[dev-dependencies.proptest]
-version = "0.9.0"
-# Disable `fork` feature to allow compiling on webassembly
-default-features = false
-features = ["std", "bit-set", "break-dead-code"]
